requirements.txt
alembic==1.13.1
alembic-postgresql-enum==1.1.1
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
astroid==3.0.1
bcrypt==4.0.1
black==24.1.1
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
dill==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
google-api-core==2.17.1
google-auth==2.28.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.18.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.62.0
greenlet==3.0.1
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httptools==0.6.1
idna==3.4
isort==5.12.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.6.1
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==3.11.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
py-xid==1.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pylint==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.2
types-passlib==1.7.7.13
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
